Move theme related things into its own subfolder
This will make contributions easier and less prone to conflicts.
