+++ src/NetscriptFunctions/UserInterface.ts
@@ -4,8 +4,8 @@ import { IPlayer } from "../PersonObjects/IPlayer";
 import { getRamCost } from "../Netscript/RamCostGenerator";
 import { GameInfo, IStyleSettings, UserInterface as IUserInterface, UserInterfaceTheme } from "../ScriptEditor/NetscriptDefinitions";
 import { Settings } from "../Settings/Settings";
-import { ThemeEvents } from "../ui/React/Theme";
-import { defaultTheme } from "../Settings/Themes";
+import { ThemeEvents } from "../Themes/ui/Theme";
+import { defaultTheme } from "../Themes/Themes";
 import { defaultStyles } from "../Settings/Styles";
 import { CONSTANTS } from "../Constants";
 import { hash } from "../hash/hash";
