+++ src/Themes/data/index.ts
@@ -1,0 +1,14 @@
+export { Theme as Default } from "./default";
+export { Theme as DefaultLite } from "./default-lite";
+export { Theme as Monokai } from "./monokai-ish";
+export { Theme as Warmer } from "./warmer";
+export { Theme as DarkPlus } from "./dark-plus";
+export { Theme as MayukaiDark } from "./mayukai-dark";
+export { Theme as Purple } from "./purple";
+export { Theme as SmoothGreen } from "./smooth-green";
+export { Theme as Dracula } from "./dracula";
+export { Theme as DarkBlue } from "./dark-blue";
+export { Theme as DiscordLike } from "./discord-like";
+export { Theme as OneDark } from "./one-dark";
+export { Theme as MutedGoldBlue } from "./muted-gold-blue";
+export { Theme as Light } from "./light";
